CRUD_user_api/requirements.txt
flask
firebase-admin
requests
google-cloud-firestore
gunicorn
python-dotenv
bcrypt